+++ src/components/AddItem/constants.ts
@@ -1,0 +1,4 @@
+export const DATA_HOOKS = {
+  ADD_ITEM: 'add-item',
+  TEXT: 'text',
+};