+++ src/components/AddItem/index.tsx
@@ -1,0 +1,7 @@
+export {
+  AddItem,
+  AddItemProps,
+  ICON_SIZE,
+  ALIGNMENT,
+  DIRECTION,
+} from './AddItem';