+++ src/components/ColorPicker/ColorPickerItem/ColorPickerItem.tsx
@@ -11,9 +11,6 @@ import { MobileTooltip } from './MobileTooltip';
 import { TPAComponentsConsumer } from '../../TPAComponentsConfig';
 import { TOOLTIP_COMMON_PROPS } from './tooltipCommonProps';
 
-const TOOLTIP_SHOW_DELAY = 0;
-const TOOLTIP_HIDE_DELAY = 1000;
-
 export interface ColorPickerItemProps extends RadioButtonProps {
   className?: string;
   key?: number;
@@ -58,14 +55,26 @@ export class ColorPickerItem extends React.Component<
           {...TOOLTIP_COMMON_PROPS}
           data-hook={props.tooltipDataHook}
           content={props.tooltip}
-          showDelay={TOOLTIP_SHOW_DELAY}
-          hideDelay={TOOLTIP_HIDE_DELAY}
         >
           {radioVisual}
         </Tooltip>
       );
     }
     return radioVisual;
+  };
+
+  getFocusOnHoverBehaviour = (isMobile: boolean) => {
+    if (isMobile) {
+      return {};
+    }
+    return {
+      onHover: () => {
+        this.setState({ focused: true });
+      },
+      onIconBlur: () => {
+        this.setState({ focused: false });
+      },
+    };
   };
 
   render = () => {
@@ -112,8 +121,7 @@ export class ColorPickerItem extends React.Component<
               mobile,
             )}
             onChange={onChange}
-            onHover={() => this.setState({ focused: true })}
-            onIconBlur={() => this.setState({ focused: false })}
+            {...this.getFocusOnHoverBehaviour(mobile)}
             name={name}
           />
         )}