+++ src/components/Tabs/Tab/Tab.tsx
@@ -51,7 +51,11 @@ export const Tab = React.forwardRef<HTMLLIElement, TabProps>((props, ref) => {
       aria-current={isActive}
     >
       <span role="link">{title}</span>
-      <div className={classes.focusIndicatorWrapper} tabIndex={-1}>
+      <div
+        className={classes.focusIndicatorWrapper}
+        tabIndex={-1}
+        aria-hidden="true"
+      >
         <div className={classes.focusIndicator} tabIndex={-1} />
       </div>
     </li>